+++ src/library/modreg/src/ksmooth.c
@@ -13,7 +13,7 @@ static double dokern(double x, int kern)
     return (0.0); /* -Wall */
 }
 
-void BDRksmooth(double *x, double *y, int *n, double *xp, double *yp, long *np, long *kern, double *bandwidth)
+void BDRksmooth(double *x, double *y, int *n, double *xp, double *yp, int *np, int *kern, double *bandwidth)
 {
     int i, imin = 0, j;
     double cutoff = 0.0, num, den, x0, w, bw = *bandwidth;
